==> affyData/markdown/02_diagnostics.Rmd <==
__Author: Nitesh Turaga, May 6th 2014__


Goal: from Affymetrix raw data stored in CEL files to differential gene expression
========================================================


### Some Basics

Getting  the working directory, in my case

```{r getwd,comment=NA,}
getwd()
```

Setting the working directory, in my case:
```{r setwd,eval=TRUE,comment=NA}
setwd("~/Documents/GeneExpressionDataAnalysis/affyData/")
````

Check the date

```{r checkDate,comment=NA}
date()
```

Clear the workspace, note the two "embedded" functions
```{r clearWorkSpace,comment=NA}
rm(list=ls())
```

Load the libraries needed for the analysis, note the use of "require"
```{r loadLibraries,comment=NA}
require(affy)
require(hgu95a.db)
```



### Load data previously saved, and check their classes

```{r loadPreviousData,comment=NA}
load("./objs/affyData.rda")
```


Check the the workspace content
```{r checkWorkspace}
ls()
```

Check classes
```{r checkClasses,comment=NA}
class(dat)

### Note: dat.rma is an ExpresssionSet, which is a 
class(dat.rma)
```

Access to data in the class
---------------------------

Slots/elements contained in the two classes can be seen using: 
Both **AffyBatch** and **ExpressionSet** are classes which **EXTENDS** the functionality
of an **Biobase::eSet**. When a class X **EXTENDS** another base class Y, it means that,
you have all the basic functionality of the parent class, and X also adds some functionality of
its own. This is very popular in object oriented programming, called Inheritance.

__NOTE: SLOT NAMES USUALLY WORK AS ACCESSORS__
```{r slotNames,comment=NA}
slotNames(dat)
slotNames(dat.rma)
```

The **exprs** extracts normalized and summarized gene expression
```{r expressionMatrix,comment=NA}
dat.expr <- exprs(dat.rma)
dim(dat.expr)
```

The **pData** function extracts the phenotypic information
```{r pData, comment=NA}
pData(dat.rma)
```
The **annotation** function extracts features information
```{r annotation,comment=NA}
annotation(dat.rma)
```

The **pm** function extracts perfect match probe intensities.
_Perfect match probes are intended to match perfectly the target sequence_.
```{r pm,comment=NA}
dat.pm <- pm(dat)
dim(dat.pm)
```

The **mm** extract mismatch probe intensities
_Mismatch probes having one base mismatch with the target sequence intended to account for non-specific binding_
```{r mm,comment=NA}
dat.mm <- mm(dat)
dim(dat.mm)
```

MA Plot 
---------

**MA-plots** are used to study dependences between the log ratio of two variables and the mean values of the same two variables. The log ratios of the two measurements are called M values (from “minus” in the log scale) and are represented in the vertical axis. The mean values of the two measurements are called A values (from “average” in the log scale) and are represented in the horizontal axis.

In microarray data contexts an MA-plot is used to compare two channels of intensity measurements. These two channels can be the red and green channels of one single chip of a two-color platform or the intensity measurements of two different arrays when using a single-channel platform. In a single-channel context at least two arrays are needed to draw a meaningful MA-plot.

$M = log_{2}(R/G) = log_{2}(R) - log_{2}(G)$

$A = \frac{1}{2}log_{2}(R*G) = \frac{1}{2}(log_{2}(R) + log_{2}(G))$



### We can now make an MA-plot

We first define a function to divide the plotting area in rows and columns, 
if you are unfamiliar with how to set the plotting area, you can use the **?par**.

```{r makePlotArea,comment=NA}
mypar <- function (nRow = 1, nCol = 1, ptsExp = 1) {
  par(mar = c(2, 2, 2, 1)) # c(bottom, left, top, right)
	par(oma = c(2, 1, 1, 1)) #  c(bottom, left, top, right) giving the size of the outer margins in lines of text.
	par(mfrow = c(nRow, nCol)) # figures are drawn with (nr * nc) matrix 
	par(cex = ptsExp) # magnification of plotting text and points
}
```


We can compute rows and columns number based on the nuber of samples.
You can do this also manually:
```{r computeRowsCols,comment=NA}
nc <- ceiling(sqrt(ncol(dat.rma)))
nr <- ceiling(ncol(dat.rma)/nc)
```

### We can make a plot for the raw data, before normalization
```{r MAplot,fig.width=15,fig.height=15,comment=NA}
mypar(nr,nc,0.5)
MAplot(dat[,],pairs=F,plot.method="smoothScatter")
```

We can save the plot to a file,
```{r saveMAPlot,comment=NA}
bitmap(file="./figs/MApl.raw.png",width=20, height=10, res = 400)
mypar(nr,nc,0.5)
par(oma=c(2,1,1,1))
MAplot(dat[,],pairs=F,plot.method="smoothScatter")
dev.off()
```


### MA-plot after normalization
```{r MAplotAFterNormalization,fig.width=15,fig.height=15,comment=NA}
mypar(nr,nc,0.5)
par(oma=c(2,1,1,1)) #overwrite initial mypar settings
a <- exprs(dat.rma)
for (i in 1:ncol(a)){
  ma.plot(A=((a[,i]+apply(a,1,median))/2),
    M=(a[,i]-apply(a,1,median)),
    show.statistics=TRUE,cex.main=1,
    span=1/3, family.loess="gaussian", cex = 0.75,
    plot.method="smoothScatter",add.loess = TRUE,
    lwd = 2, lty = 2, loess.col = "red",ylim=c(-6,6),
    main=paste(colnames(a)[i],
	  '\n vs pseudo-median reference chip'))
}
```


MA-plot after normalization: saved to a file
```{r saveMAPlotAfterNormalization,comment=NA}
bitmap(file='./figs/MApl.rma.png',width=20,height=10,res=400)
mypar(nr,nc,0.5)
par(oma=c(2,1,1,1))
a <- exprs(dat.rma)
for (i in 1:ncol(a)){
  ma.plot(A=((a[,i]+apply(a,1,median))/2),
    M=(a[,i]-apply(a,1,median)),
    show.statistics=TRUE,cex.main=1,
    span=1/3, family.loess="gaussian", cex = 0.75,
    plot.method="smoothScatter",add.loess = TRUE,
    lwd = 2, lty = 2, loess.col = "red",ylim=c(-6,6),
    main=paste(colnames(a)[i],
	  'vs pseudo-median reference chip'))
}
dev.off()
```


### Diagnostic Plots: 2D-Image

```{r bitMap,fig.width = 10,fig.height=10,comment=NA}
bitmap(file="./figs/ima2D.png",width=2*nr, height=2*nc, res = 144)
mypar(nr,nc,1) ; par(mar=c(1,1,1,1))
image(dat[,])
dev.off()
```

```{r affy_image_demo,echo=FALSE,comment=NA}
mypar(nr,nc,1) ; par(mar=c(1,1,1,1))
image(dat[,])

```

### Log2 intensities boxplots: before and after normalization
```{r boxPlots,comment=NA}
bitmap(file="./figs/exprBoxplot.png",width=15, height=15, res = 144)
mypar(2,1,0.75)
par(mar=c(10,1,1,1))
### Before
boxplot(log2(exprs(dat))~col(exprs(dat)),
	names=pData(dat.rma)[,"Row.names"],
	las=2,col='blue')
### After
boxplot(exprs(dat.rma)~col(exprs(dat.rma)),
	names=pData(dat.rma)[,"Row.names"],
	las=2,col='blue')
### Close the device
dev.off()
```

```{r boxplots_for_demo_inClass,comment=NA,echo=FALSE}
mypar(2,1,0.75)
par(mar=c(10,1,1,1))
### Before
boxplot(log2(exprs(dat))~col(exprs(dat)),
  names=pData(dat.rma)[,"Row.names"],
	las=2,col='blue')
### After
boxplot(exprs(dat.rma)~col(exprs(dat.rma)),
	names=pData(dat.rma)[,"Row.names"],
	las=2,col='blue')
```


### RNA Degradation

The functions AffyRNAdeg, summaryAffyRNAdeg, and plotAffyRNAdeg aid in assessment of RNA quality. Since RNA degradation typically starts from the $5'$ end of the molecule, we would expect probe intensities to be systematically lowered at that end of a probeset when compared to the $3'$ end.

Compute RNA degradation
```{r RNAdegradation,comment=NA}
deg <- AffyRNAdeg(dat)

names(deg) # We plot the degradation and slope
```


Make RNA degradation plot
```{r makeRNADegradationPlot,comment=NA}
bitmap(file="./figs/degRNA.png",width=5, height=5, res = 1000)
mypar(2,2,0.75)
plotAffyRNAdeg(deg,cols=c(1:ncol(dat.rma)))
plot(density(deg$slope),main="Slope")
boxplot(deg$slope,main="Slope")
### Close the device
dev.off()
```

```{r rnaDegradationDemo,comment=NA,echo=FALSE}
mypar(2,2,0.75)
plotAffyRNAdeg(deg,cols=c(1:ncol(dat.rma)))
plot(density(deg$slope),main="Slope")
boxplot(deg$slope,main="Slope")
```

Find the scan date
```{r dateOfScan,comment=NA}
dateOfScan <- protocolData(dat.rma)@data$ScanDate
#Don't need the time, so we will remove that using regular expressions
dateOfScan <- gsub(" .+", "", dateOfScan)
boxplot(log2(exprs(dat)), col=1+as.numeric(dateOfScan))
```

### Session information
```{r sessionInformation,comment=NA}
sessionInfo()
```
